+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file. This projec
 to [Semantic Versioning] (http://semver.org/). For change log format,
 use [Keep a Changelog] (http://keepachangelog.com/).
 
+## [1.0.0-beta2] - In progress
+
+### Added
+
+- `SchemaContainerInterface::hasColumn()` method
+- `SchemaContainerInterface::getColumn()` method
+
+### Removed
+
+- Dependency with reflections
+- @package attributes from PhpDoc
+
 ## [1.0.0-beta1] - 2021-06-02
 
 Initial development.